Moves enum to new enums module
